requirements.txt
pandas==1.5.3
langchain==0.1.12
neo4j==5.14.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.41.0